decouple rejection handler from echo_x_request_id
