+++ src/routes/mod.rs
@@ -10,3 +10,6 @@ pub use points::points;
 
 mod lines_batch;
 pub use lines_batch::lines_batch;
+
+mod custom_rejection_handler;
+pub use custom_rejection_handler::custom_rejection_handler;
